Lambda function to process form submissions
Much of the functionality of other scripts on this repo have been jammed in here because I was having a hard time deploying multiple files to a lambda container. 

This is currently hosted on lambda, and receives a json payload created from the index.html page in this repo. It processes the form and makes corresponding records in Airtable, and also generates a pdf and mails it to an internal email.